RewriteTools

Documentation for RewriteTools.

RewriteTools.Term

RewriteTools.flatten_term

RewriteTools.rewrite_rhs

RewriteTools.@capture

RewriteTools.@rule

RewriteTools.@slots

RewriteTools.Term — Type

Term(f, args::AbstractArray)

Symbolic expression representing the result of calling f(args...).

operation(t::Term) returns f

arguments(t::Term) returns args

RewriteTools.flatten_term — Method

flatten_term(⋆, x)

Return a flattened expression with the numbers at the back.

Example

julia> ex = RewriteTools.flatten_term(+, term(+, :a, term(+, :b, :c)))
a + b + c

julia> ex == term(+, :a, :b, :c)
true

RewriteTools.rewrite_rhs — Method

rewrite_rhs(expr::Expr)

Rewrite the expr by dealing with :where if necessary. The :where is rewritten from, for example, ~x where f(~x) to f(~x) ? ~x : nothing.

RewriteTools.@capture — Macro

@capture [SLOTS...] ex pattern

Uses a Rule object to capture an expression if it matches the pattern. Returns true and injects slot variable match results into the calling scope when the pattern matches, otherwise returns false. The rule language for specifying the pattern is the same in @capture as it is in @rule.

julia> @syms a; ex = a^a;

julia> if @capture ex (~x)^(~x)
           @show x
       elseif @capture ex 2(~y)
           @show y
       end;
x = a

See also: @rule, @slots

RewriteTools.@rule — Macro

@rule [SLOTS...] LHS => RHS

Creates a Rule object. A rule object is callable, and takes an expression and rewrites it if it matches the LHS pattern to the RHS expression, returns nothing otherwise. The rule language is described below.

LHS can be any possibly nested function call expression where any of the arguments can optionally be a Slot (~x) or a Segment (~x...) (described below).

SLOTS is an optional list of symbols to be interpted as slots or segments directly (without using ~). To declare slots for several rules at once, see the @slots macro.

If an expression matches LHS entirely, then it is rewritten to the result of the expression RHS, whose local scope includes the slot matches as variables.

Slot:

A Slot variable is written as ~x and matches a single expression. x is the name of the variable. If a slot appears more than once in an LHS expression then expression matched at every such location must be equal (as shown by isequal).

Example:

Simple rule to turn any sin into cos:

julia> @syms a b c
(a, b, c)

julia> r = @rule sin(~x) => cos(x)
sin(~x) => cos(x)

julia> r(sin(1+a))
cos((1 + a))

A rule with 2 segment variables

julia> r = @rule sin(~x + ~y) => sin(x)*cos(y) + cos(x)*sin(y)
sin(~x + ~y) => sin(x) * cos(y) + cos(x) * sin(y)

julia> r(sin(a + b))
cos(a)*sin(b) + sin(a)*cos(b)

A rule that matches two of the same expressions:

julia> r = @rule sin(~x)^2 + cos(~x)^2 => 1
sin(~x) ^ 2 + cos(~x) ^ 2 => 1

julia> r(sin(2a)^2 + cos(2a)^2)
1

julia> r(sin(2a)^2 + cos(a)^2)
# nothing

A rule without ~

julia> r = @slots x y z @rule x(y + z) => xy + xz
x(y + z) => xy + xz

Segment:

A Segment variable matches zero or more expressions in the function call. Segments may be written by splatting slot variables (~x...).

Example:

This implements the distributive property of multiplication: +(~ys...) matches expressions like a + b, a+b+c and so on. On the RHS ys presents as any old julia array.

julia> r = @rule ~x * +((~ys...)) => sum(map(y-> x * y, ys));

julia> r(2 * (a+b+c))
((2 * a) + (2 * b) + (2 * c))

A segment without ~.

julia> r = @slots xs @rule min(xs...) => foldl(min, xs, Inf);

julia> r(min(a, b, c))
min(min(a, b), c)

**Predicates**:

There are two kinds of predicates, namely over slot variables and over the whole rule.
For the former, predicates can be used on both `~x` and `~x...` like `~x::f` or `~x::f...`.
Here `f` can be any julia function. In the case of a slot the function gets a single
matched subexpression, in the case of segment, it gets an array of matched expressions.

The predicate should return `true` if the current match is acceptable, and `false`
otherwise.

julia julia> twoπs(x::Number) = abs(round(x/(2π)) - x/(2π)) < 10^-9 twoπs (generic function with 1 method)

julia> twoπs(x) = false twoπs (generic function with 2 methods)

julia> r = @slots x y z @rule sin(+(x..., y::twoπs, z...)) => sin(+(x..., z...)) sin(+(x..., y::twoπs, z...)) => sin(+(x..., z...))

julia> r(sin(a+3π))

julia> r(sin(a+6π)) sin(a)

julia> r(sin(a+6π+c)) sin((a + c))

For the predicate over the whole rule, use `@rule <LHS> => <RHS> where <predicate>`:

julia> @syms a b;

julia> predicate(x) = x === a;

julia> r = @rule ~x => x where f(x);

julia> r(a) a

julia> r(b) === nothing true ```

Note that this is syntactic sugar and that it is the same as something like @rule ~x => f(x) ? x : nothing.

Compatibility: Segment variables may still be written as (~~x), and slot (~x) and segment (~x... or ~~x) syntaxes on the RHS will still substitute the result of the matches.

See also: @capture, @slots

RewriteTools.@slots — Macro

@slots [SLOTS...] ex

Declare SLOTS as slot variables for all @rule or @capture invocations in the expression ex.

Example:

julia> @slots x y z a b c Chain([
    (@rule x^2 + 2x*y + y^2 => (x + y)^2),
    (@rule x^a * y^b => (x*y)^a * y^(b-a)),
    (@rule +(x...) => sum(x)),
])
Chain(Rule{Term[...]))

See also: @rule, @capture